Git: Ignore built files
This commit will make git ignore built files

Hugo is a static site generator. That means it builds every file in to
markdown. However, it needs to a directory to store the built files.
These files are stored in the public directory in the site root.
This commit will cause git to ignore it.