Exercise: Clean up title
This commit will remove the "Tcpl" in the title

I felt that appending "Tcpl" to title was rather dumb. After all, the
visitors of this site know that I attempt "The C Programming Language"!
Hell, I have got it written in Bold as the title in the homepage!

So, I decided to remove Tcpl from the title. A quick sed command did the
trick (which iterated over find) did the trick.