Archtypes: Create an archtype for exercises
This commit will create an archtype exercises, which can then be
used to create new exercises easily

Hugo archtypes can be used as a template to create a post. I am simply
making use of this feature so that I can write the solution to a
question more easily.

Useful link:

- https://gohugo.io/content-management/archetypes/