All Exercises: Update makefile cleaning mechanism
This commit will cause make to iterate over `$(objects)` for making
`all` and `clean`. It will also make clean a `.PHONY` target. This will also
make the previously written makefiles of the same "style" as of the
archtypes' makefile.

I started to feel that they way I used to clean files previously in
the makefile was very error prone, and was very manual in nature.

So I decided to write a better mechanism, where I iterated over a
variable `objects` in my targets, `all` and `clean`. This makes sure
that all and clean are synced.

Another thing I did was make clean a .PHONY target, and used `-rm`
instead of `rm`, You can find the explanation in the following links:

- https://www.gnu.org/software/make/manual/html_node/Phony-Targets.html
- https://www.gnu.org/software/make/manual/html_node/Cleanup.html

Another useful link:

- https://www.gnu.org/software/make/manual/html_node/Using-Variables.html